consolidated accessibility services into one class and added 2 nav buttons to overlay
