+++ app/src/main/AndroidManifest.xml
@@ -28,38 +28,19 @@
 
         <service
             android:name=".OverlayService"
-            android:label="@string/service_name"
-            android:enabled="true"
-            android:exported="false">
-            <intent-filter>
-                <action android:name="android.accessibilityservice.AccessibilityService" />
-            </intent-filter>
-            <meta-data
-                android:name="android.accessibilityservice"
-                android:resource="@xml/accessibility_service_config" />
-        </service>
-
-
-        <!--        <service-->
-<!--            android:name=".OverlayService"-->
-<!--            android:configChanges="orientation"-->
-<!--            android:exported="true"-->
-<!--            android:screenOrientation="fullSensor">-->
-<!--        </service>-->
-
-        <service
-            android:name=".MyAccessibilityService"
             android:enabled="true"
             android:exported="false"
             android:label="@string/accessibility_service_label"
+            android:description="@string/accessibility_service_description"
             android:permission="android.permission.BIND_ACCESSIBILITY_SERVICE">
+
             <intent-filter>
                 <action android:name="android.accessibilityservice.AccessibilityService" />
             </intent-filter>
 
             <meta-data
                 android:name="android.accessibilityservice"
-                android:resource="@xml/accessibility_service_config" />
+                android:resource="@xml/overlay_service_config" />
         </service>
     </application>
