+++ app/src/main/res/layout/activity_overlay.xml
@@ -10,12 +10,12 @@
         android:id="@+id/sidebar"
         android:layout_width="match_parent"
         android:layout_height="360dp"
-        android:text="recent" />
+        android:text="recentscreen" />
 
     <Button
         android:id="@+id/sidebarbottom"
         android:layout_width="match_parent"
         android:layout_height="360dp"
         android:layout_marginTop="360dp"
-        android:text="back" />
+        android:text="backbtn" />
 </FrameLayout>
